RENAME: api type을 auth와 todo type으로 이동
