+++ src/types/todo.ts
@@ -8,3 +8,21 @@ export interface ITodo {
 
 // 전체 Todo
 export type ITodoList = ITodo[];
+
+// apis
+// 2-1 createTodo
+export interface ICreateTodoRequeset {
+  todo: string;
+}
+
+// 2-3 updateTodo
+export interface IUpdateTodoRequest {
+  id: number;
+  todo: string;
+  isCompleted: boolean;
+}
+
+// 2-4 deleteTodo
+export interface IDeleteTodoRequest {
+  id: number;
+}
